+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - **[Fix]** Fix path to source in source maps.
+- **[Fix]** Disable `number-literal-format` in default Tslint rules. It enforced uppercase for hex.
 
 ## 0.15.5 (2017-11-10)