+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **[Fix]** Fix path to source in source maps.
 - **[Fix]** Disable `number-literal-format` in default Tslint rules. It enforced uppercase for hex.
+- **[Internal]** Enable integration with Greenkeeper.
 
 ## 0.15.5 (2017-11-10)